+++ packages/core/test/lib/integrations/dedupe.test.ts
@@ -1,5 +1,4 @@
-import { expect } from 'chai';
-import { Dedupe } from '../../src/integrations/dedupe';
+import { Dedupe } from '../../../src/integrations/dedupe';
 
 function clone<T>(data: T): T {
   return JSON.parse(JSON.stringify(data));
@@ -58,21 +57,21 @@ describe('Dedupe', () => {
   describe('shouldDropEvent(messageEvent)', () => {
     it('should not drop if there was no previous event', () => {
       const event = clone(messageEvent);
-      expect(dedupe.shouldDropEvent(event)).equal(false);
+      expect(dedupe.shouldDropEvent(event)).toBe(false);
     });
 
     it('should not drop if events have different messages', () => {
       const eventA = clone(messageEvent);
       const eventB = clone(messageEvent);
       eventB.message = 'EvilMorty';
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
     });
 
     it('should not drop if events have same messages, but different stacktraces', () => {
       const eventA = clone(messageEvent);
       const eventB = clone(messageEvent);
       eventB.stacktrace.frames[0].colno = 1337;
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
     });
 
     it('should drop if there are two events with same messages and no fingerprints', () => {
@@ -80,13 +79,13 @@ describe('Dedupe', () => {
       delete eventA.fingerprint;
       const eventB = clone(messageEvent);
       delete eventB.fingerprint;
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(true);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(true);
     });
 
     it('should drop if there are two events with same messages and same fingerprints', () => {
       const eventA = clone(messageEvent);
       const eventB = clone(messageEvent);
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(true);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(true);
     });
 
     it('should not drop if there are two events with same message but different fingerprints', () => {
@@ -95,42 +94,42 @@ describe('Dedupe', () => {
       eventA.fingerprint = ['Birdperson'];
       const eventC = clone(messageEvent);
       delete eventC.fingerprint;
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
-      expect(dedupe.shouldDropEvent(eventA, eventC)).equal(false);
-      expect(dedupe.shouldDropEvent(eventB, eventC)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
+      expect(dedupe.shouldDropEvent(eventA, eventC)).toBe(false);
+      expect(dedupe.shouldDropEvent(eventB, eventC)).toBe(false);
     });
   });
 
   describe('shouldDropEvent(exceptionEvent)', () => {
     it('should not drop if there was no previous event', () => {
       const event = clone(exceptionEvent);
-      expect(dedupe.shouldDropEvent(event)).equal(false);
+      expect(dedupe.shouldDropEvent(event)).toBe(false);
     });
 
     it('should drop when events type, value and stacktrace are the same', () => {
       const event = clone(exceptionEvent);
-      expect(dedupe.shouldDropEvent(event, event)).equal(true);
+      expect(dedupe.shouldDropEvent(event, event)).toBe(true);
     });
 
     it('should not drop if types are different', () => {
       const eventA = clone(exceptionEvent);
       const eventB = clone(exceptionEvent);
       eventB.exception.values[0].type = 'TypeError';
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
     });
 
     it('should not drop if values are different', () => {
       const eventA = clone(exceptionEvent);
       const eventB = clone(exceptionEvent);
       eventB.exception.values[0].value = 'Expected number, got string';
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
     });
 
     it('should not drop if stacktraces are different', () => {
       const eventA = clone(exceptionEvent);
       const eventB = clone(exceptionEvent);
       eventB.exception.values[0].stacktrace.frames[0].colno = 1337;
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
     });
 
     it('should drop if there are two events with same exception and no fingerprints', () => {
@@ -138,13 +137,13 @@ describe('Dedupe', () => {
       delete eventA.fingerprint;
       const eventB = clone(exceptionEvent);
       delete eventB.fingerprint;
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(true);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(true);
     });
 
     it('should drop if there are two events with same exception and same fingerprints', () => {
       const eventA = clone(exceptionEvent);
       const eventB = clone(exceptionEvent);
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(true);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(true);
     });
 
     it('should not drop if there are two events with same exception but different fingerprints', () => {
@@ -153,9 +152,9 @@ describe('Dedupe', () => {
       eventA.fingerprint = ['Birdperson'];
       const eventC = clone(exceptionEvent);
       delete eventC.fingerprint;
-      expect(dedupe.shouldDropEvent(eventA, eventB)).equal(false);
-      expect(dedupe.shouldDropEvent(eventA, eventC)).equal(false);
-      expect(dedupe.shouldDropEvent(eventB, eventC)).equal(false);
+      expect(dedupe.shouldDropEvent(eventA, eventB)).toBe(false);
+      expect(dedupe.shouldDropEvent(eventA, eventC)).toBe(false);
+      expect(dedupe.shouldDropEvent(eventB, eventC)).toBe(false);
     });
   });
 });